feat: implement import by path to extend orchestrators
